download data sets from Statistics Denmark

You can use this function to retrieve data sets from Statistics Denmark. It also let you choose which variables to return.

dst_download(tableID, vars, lang = "en")

Arguments

tableID
the data set's ID. See dst_tables() to get the ID.

vars
the variables to return. If not specified it will return time and value of the what the data set measure.

lang
used language in the data set. Can take the values "da" for danish and "en" for english

Examples

dst_download("folk1a")

#> Warning: No vars is specified. Only time will be included

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   TID = col_character(),
#>   INDHOLD = col_integer()
#> )

#> # A tibble: 38 x 2
#>       TID INDHOLD
#>     <chr>   <int>
#>  1 2008Q1 5475791
#>  2 2008Q2 5482266
#>  3 2008Q3 5489022
#>  4 2008Q4 5505995
#>  5 2009Q1 5511451
#>  6 2009Q2 5515287
#>  7 2009Q3 5519441
#>  8 2009Q4 5532531
#>  9 2010Q1 5534738
#> 10 2010Q2 5540241
#> # ... with 28 more rows

dst_download("folk1a", c("ALDER", "CIVILSTAND"))

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   TID = col_character(),
#>   ALDER = col_character(),
#>   CIVILSTAND = col_character(),
#>   INDHOLD = col_integer()
#> )

#> # A tibble: 24,130 x 4
#>       TID  ALDER        CIVILSTAND INDHOLD
#>     <chr>  <chr>             <chr>   <int>
#>  1 2008Q1  Total             Total 5475791
#>  2 2008Q1  Total     Never married 2552700
#>  3 2008Q1  Total Married/separated 2176977
#>  4 2008Q1  Total           Widowed  317250
#>  5 2008Q1  Total          Divorced  428864
#>  6 2008Q1 0 year             Total   64412
#>  7 2008Q1 0 year     Never married   64412
#>  8 2008Q1 0 year Married/separated       0
#>  9 2008Q1 0 year           Widowed       0
#> 10 2008Q1 0 year          Divorced       0
#> # ... with 24,120 more rows