Retrieve subjects from Statistics Denmark

You can use `dst_subject()` to get the subjects, which Statistics Denmark use to categorize their data sets.

dst_subjects(lang = "en", columns = c("id", "description"))

Arguments

lang
description language - "da" for danish and "en" for english.

columns
columns to return. Can take the values "id", "description", "active", "hasSubject" or "subjects".

Value

Tibble containing the specified columns in the chosen language.

Examples

dst_subjects()

#> # A tibble: 12 x 2
#>       id                               description
#>  * <chr>                                     <chr>
#>  1    02                  Population and elections
#>  2    05                         Living conditions
#>  3    03                   Education and knowledge
#>  4    18               Culture and National Church
#>  5    04                 Labour, income and wealth
#>  6    06                    Prices and consumption
#>  7    14 National accounts and government finances
#>  8    16                   Money and credit market
#>  9    13                          External economy
#> 10    07                Business sector in general
#> 11    11                          Business sectors
#> 12    01         Geography, environment and energy

dst_subjects(columns = c("id", "description", "active")) # is the data set active

#> # A tibble: 12 x 3
#>       id                               description active
#>  * <chr>                                     <chr>  <lgl>
#>  1    02                  Population and elections   TRUE
#>  2    05                         Living conditions   TRUE
#>  3    03                   Education and knowledge   TRUE
#>  4    18               Culture and National Church   TRUE
#>  5    04                 Labour, income and wealth   TRUE
#>  6    06                    Prices and consumption   TRUE
#>  7    14 National accounts and government finances   TRUE
#>  8    16                   Money and credit market   TRUE
#>  9    13                          External economy   TRUE
#> 10    07                Business sector in general   TRUE
#> 11    11                          Business sectors   TRUE
#> 12    01         Geography, environment and energy   TRUE