retrieve tables from Statistics Denmark

You can use `dst_tables()` to get a tibble, which contain information about the data sets available at the Bureau.

dst_tables(subjectsID, lang = "en", columns = c("id", "text"))

Arguments

subjectsID
Should the data sets be within a specific field. See dst_subject() for IDs.

lang
the desired language. Can take the values "da" for danish and "en" english.

columns
the returned columns. Can take the values "id", "text", "unit", "updated", "firstPeriod", "latestPeriod", "active" or "variables"

Value

Tibble which contain information about all tables available at Danish Statistics. If subject is provided only tables within the subject is returned. tibble will contain the columns specified in the columns argument. The language is as default English otherwise "da" is provided in the lang argument.

Examples

dst_tables() # retrieve all tables

#> # A tibble: 1,847 x 2
#>        id                                                          text
#>  *  <chr>                                                         <chr>
#>  1 FOLK1A                    Population at the first day of the quarter
#>  2 FOLK1B                    Population at the first day of the quarter
#>  3 FOLK1C                    Population at the first day of the quarter
#>  4 FOLK1D                    Population at the first day of the quarter
#>  5 FOLK1E                    Population at the first day of the quarter
#>  6  FOLK2                                         Population 1. January
#>  7  FOLK3                                         Population 1. January
#>  8     FT                          Population figures from the censuses
#>  9  BEF5F People born in Faroe Islands and living in Denmark 1. January
#> 10  BEF5G     People born in Greenland and living in Denmark 1. January
#> # ... with 1,837 more rows

dst_tables(subjectsID = "02") # Population and elections data sets

#> # A tibble: 153 x 2
#>        id                                                          text
#>  *  <chr>                                                         <chr>
#>  1 FOLK1A                    Population at the first day of the quarter
#>  2 FOLK1B                    Population at the first day of the quarter
#>  3 FOLK1C                    Population at the first day of the quarter
#>  4 FOLK1D                    Population at the first day of the quarter
#>  5 FOLK1E                    Population at the first day of the quarter
#>  6  FOLK2                                         Population 1. January
#>  7  FOLK3                                         Population 1. January
#>  8     FT                          Population figures from the censuses
#>  9  BEF5F People born in Faroe Islands and living in Denmark 1. January
#> 10  BEF5G     People born in Greenland and living in Denmark 1. January
#> # ... with 143 more rows