get data set variables

You can use this function to get a tibble which has information about the available variables in the respective data set as well as an description of those

dst_variables(tableID, lang = "en", columns = c("id", "text"))

Arguments

tableID
the id of the desired table. See dst_tables() to get an overview of the table IDs.

lang
the language used in the description of the variables. "en" for english "da" for danish.

columns
the columns returned. columns can take the following values: id, text, elimination, time or map

Examples

dst_variables("folk1a")

#>           id           text
#> 1     OMRÅDE         region
#> 2        KØN            sex
#> 3      ALDER            age
#> 4 CIVILSTAND marital status
#> 5        Tid           time

dst_variables("folk1a", columns = c("id", "text", "time"))

#>           id           text  time
#> 1     OMRÅDE         region FALSE
#> 2        KØN            sex FALSE
#> 3      ALDER            age FALSE
#> 4 CIVILSTAND marital status FALSE
#> 5        Tid           time  TRUE